PS/2: honour overflow and sign correctly and purge custom rotate functions.

+++ src/ps2mouse.h
@@ -36,7 +36,7 @@ uint8_t errcnt;
 uint8_t g_mouse_enabled; ///< >0 if mouse/trackpoint is active and reports shall be generated
 uint8_t g_ps2_connected; ///< >0 if a PS/2 device was detected.
 
-void ps2_read_mouse(int *dx, int *dy, uint8_t *BTNS );
+void ps2_read_mouse(int8_t *dx, int8_t *dy, int8_t *BTNS );
 bool ps2_init_mouse(void);
 
 uint8_t getMouseReport(USB_WheelMouseReport_Data_t *report_data);
